# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Connect API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Room {
  id: string;
  name: string;
  isPrivate: boolean;
  isGlobal: boolean;
  memberCount: number;
  messageCount: number;
  createdAt: string;
  /** @nullable */
  lastActivityAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  inactivityDeleteMinutes?: number | null;
  /** @nullable */
  emptyDeleteMinutes?: number | null;
  /** @nullable */
  messageRetentionHours?: number | null;
  /** @nullable */
  inviteCode?: string | null;
}

export interface RoomInput {
  name: string;
  inactivityDeleteMinutes?: number;
  emptyDeleteMinutes?: number;
  messageRetentionHours?: number;
}

export interface Message {
  id: string;
  roomId: string;
  userId: string;
  username: string;
  content: string;
  isAi: boolean;
  createdAt: string;
}

export interface OnlineUser {
  userId: string;
  username: string;
  isOnline: boolean;
  joinedAt: string;
}

export interface AiMessage {
  role: string;
  content: string;
}

export interface AiChatInput {
  messages: AiMessage[];
  /** @nullable */
  systemPrompt?: string | null;
}

export interface GlobalStats {
  totalOnlineUsers: number;
  totalRooms: number;
  totalMessages: number;
}

export interface ErrorResponse {
  error: string;
}