# lib/api-zod/src/generated/types/aiChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Connect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiMessage } from './aiMessage';

export interface AiChatInput {
  messages: AiMessage[];
  /** @nullable */
  systemPrompt?: string | null;
}