# lib/api-zod/src/generated/types/globalStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Connect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GlobalStats {
  totalOnlineUsers: number;
  totalRooms: number;
  totalMessages: number;
}