# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Connect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: string;
  roomId: string;
  userId: string;
  username: string;
  content: string;
  isAi: boolean;
  createdAt: Date;
}