# lib/api-zod/src/generated/types/onlineUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Connect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OnlineUser {
  userId: string;
  username: string;
  isOnline: boolean;
  joinedAt: Date;
}