# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Connect API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: string;
  name: string;
  isPrivate: boolean;
  isGlobal: boolean;
  memberCount: number;
  messageCount: number;
  createdAt: Date;
  /** @nullable */
  lastActivityAt?: Date | null;
  /** @nullable */
  expiresAt?: Date | null;
  /** @nullable */
  inactivityDeleteMinutes?: number | null;
  /** @nullable */
  emptyDeleteMinutes?: number | null;
  /** @nullable */
  messageRetentionHours?: number | null;
  /** @nullable */
  inviteCode?: string | null;
}